// bbsweb/src/app/pages/data/permissions.enum.ts
export enum Permissions {
    EDIT_ROLE = 'EDIT_ROLE',
    MANAGE_PROJECTS = 'MANAGE_PROJECTS',
    VIEW_CLIENT_DASHBOARD = 'VIEW_CLIENT_DASHBOARD',
    VIEW_RECEIPTS = 'VIEW_RECEIPTS',
    VIEW_TECHNICAL_WORKS = 'VIEW_TECHNICAL_WORKS',
    VIEW_PROGRESS = 'VIEW_PROGRESS',
    CREATE_PRIVILEGE = 'CREATE_PRIVILEGE',
    DELETE_CLIENT = 'DELETE_CLIENT',
    EDIT_PRIVILEGE = 'EDIT_PRIVILEGE',
    VIEW_ACTIVITY_FEED = 'VIEW_ACTIVITY_FEED',
    DELETE_FILES = 'DELETE_FILES',
    VIEW_PROJECT_DASHBOARD = 'VIEW_PROJECT_DASHBOARD',
    UPDATE_FILES = 'UPDATE_FILES',
    CREATE_PRELIMINARY = 'CREATE_PRELIMINARY',
    VIEW_FILES = 'VIEW_FILES',
    VIEW_CLIENTS = 'VIEW_CLIENTS',
    ONBOARD_CLIENT = 'ONBOARD_CLIENT',
    CREATE_ROLE = 'CREATE_ROLE',
    EDIT_CLIENT = 'EDIT_CLIENT',
    VIEW_PRELIMINARY = 'VIEW_PRELIMINARY',
    EDIT_PROFILE = 'EDIT_PROFILE',
    VIEW_PRIVILEGES = 'VIEW_PRIVILEGES',
    RECONCILE_INVOICES = 'RECONCILE_INVOICES',
    DELETE_PRIVILEGE = 'DELETE_PRIVILEGE',
    UPLOAD_DOCUMENTS = 'UPLOAD_DOCUMENTS',
    DELETE_ROLE = 'DELETE_ROLE',
    DELETE_USER = 'DELETE_USER',
    UPLOAD_FILES = 'UPLOAD_FILES',
    VIEW_INVOICES = 'VIEW_INVOICES',
    CREATE_USER = 'CREATE_USER',
    VIEW_MAIN_DASHBOARD = 'VIEW_MAIN_DASHBOARD',
    VIEW_REPORTS = 'VIEW_REPORTS',
    VIEW_ROLES = 'VIEW_ROLES',
    VIEW_SALES_DASHBOARD = 'VIEW_SALES_DASHBOARD',
    CREATE_INVOICES = 'CREATE_INVOICES',
    UPLOAD_TECHNICAL_WORKS = 'UPLOAD_TECHNICAL_WORKS',
    UPLOAD_BOQ = 'UPLOAD_BOQ',
    VIEW_AGENT_DASHBOARD = 'VIEW_AGENT_DASHBOARD',
    VIEW_TECHNICAL_DASHBOARD = 'VIEW_TECHNICAL_DASHBOARD',
    VIEW_LEADS = 'VIEW_LEADS',
    MANAGE_CLIENTS = 'MANAGE_CLIENTS',
    MANAGE_USERS = 'MANAGE_USERS',
    MANAGE_ROLES_PRIVILEGES = 'MANAGE_ROLES_PRIVILEGES',
    MANAGE_CLIENT_DETAILS = 'MANAGE_CLIENT_DETAILS',
    MANAGE_INVOICES = 'MANAGE_INVOICES',
    MANAGE_RECEIPTS = 'MANAGE_RECEIPTS',
    MANAGE_CLIENT_ACTIVITY = 'MANAGE_CLIENT_ACTIVITY',
    APPROVE_INVOICE = 'APPROVE_INVOICE',
    SUBMIT_SITE_VISIT_REPORT = 'SUBMIT_SITE_VISIT_REPORT',
    MANAGE_PRELIMINARY = 'MANAGE_PRELIMINARY',
    APPROVE_TECHNICAL_WORKS = 'APPROVE_TECHNICAL_WORKS',
    VIEW_CLIENT_PROFILE = 'VIEW_CLIENT_PROFILE',
    APPROVE_PRELIMINARY = 'APPROVE_PRELIMINARY',
    MANAGE_LICENCES = 'MANAGE_LICENCES',
    VIEW_LICENSE = 'VIEW_LICENSE',
    MANAGE_CLIENT_DOCUMENTS='MANAGE_CLIENT_DOCUMENTS',
    MANAGE_COUNTY_APPROVAL = 'MANAGE_COUNTY_APPROVAL',
}